# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RTN Manga API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /remote/import:
    post:
      operationId: remoteImport
      tags: [remote]
      summary: Import pages from remote URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoteImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteImportResult"
  /remote/preview:
    post:
      operationId: remotePreview
      tags: [remote]
      summary: Preview pages from remote URL before import
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemotePreviewInput"
      responses:
        "200":
          description: Preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemotePreviewResult"
...
components:
  schemas:
...
    RemoteImportInput:
      type: object
      required: [url, mangaId, chapterId]
      properties:
        url: { type: string }
        mangaId: { type: integer }
        chapterId: { type: integer }
    RemoteImportResult:
      type: object
      required: [success, importedPages]
      properties:
        success: { type: boolean }
        importedPages: { type: integer }
        message: { type: ["string", "null"] }
        errors: { type: array, items: { type: string } }
    RemotePreviewInput:
      type: object
      required: [url]
      properties:
        url: { type: string }
    RemotePreviewResult:
      type: object
      required: [title, pageCount, pageUrls]
      properties:
        title: { type: string }
        pageCount: { type: integer }
        pageUrls:
          type: array
          items: { type: string }
        chapterNumber: { type: ["number", "null"] }
        mangaTitle: { type: ["string", "null"] }
...