# lib/api-zod/src/generated/types/remoteImportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RTN Manga API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RemoteImportInput {
  url: string;
  mangaId: number;
  chapterId: number;
}